Rename local to custom
