+++ docs/step-by-step-guide.md
@@ -11,7 +11,7 @@ This blueprint consists of three independent, yet complementary, components:
 
 2. **Dataset maker app** 📂🎤: Another UI app that enables you to easily and quickly create your own Speech-to-Text dataset.
 
-3. **Finetuning script** 🛠️🤖: A script to finetune your own STT model, either using Common Voice data or your own local data created by the Dataset maker app.
+3. **Finetuning script** 🛠️🤖: A script to finetune your own STT model, either using Common Voice data or your own custom data created by the Dataset maker app.
 
 ---
 
@@ -33,24 +33,24 @@ Start by initially testing the quality of the Speech-to-Text models available in
 2. Select or add the HF model id of your choice
 3. Record a sample of your voice and get the transcribed text back. You may find that there are sometimes inaccuracies for your voice/accent/chosen language, indicating the model could benefit form finetuning on additional data.
 
-### Step 2 - Make your Local Dataset for STT finetuning
+### Step 2 - Make your Custom Dataset for STT finetuning
 
-1. Create your own, local dataset by running this command and following the instructions:
+1. Create your own, custom dataset by running this command and following the instructions:
 
     ```bash
-    python src/speech_to_text_finetune/make_local_dataset_app.py
+    python src/speech_to_text_finetune/make_custom_dataset_app.py
     ```
 
 2. Follow the instruction in the app to create at least 10 audio samples, which will be saved locally.
 
-### Step 3 - Creating a finetuned STT model using your local data
+### Step 3 - Creating a finetuned STT model using your custom data
 
-1. Configure `config.yaml` with the model, local data directory and hyperparameters of your choice. Note that if you select `push_to_hub: True` you need to have an HF account and log in locally. For example:
+1. Configure `config.yaml` with the model, custom data directory and hyperparameters of your choice. Note that if you select `push_to_hub: True` you need to have an HF account and log in locally. For example:
 
     ```bash
     model_id: openai/whisper-tiny
     dataset_id: example_data/custom
-    dataset_source: local
+    dataset_source: custom
     language: English
     repo_name: default
